Extract reusable code from ExpressionFuzzer
