+++ velox/expression/tests/ExpressionVerifier.cpp
@@ -17,6 +17,7 @@
 #include "velox/expression/tests/ExpressionVerifier.h"
 #include "velox/common/base/Fs.h"
 #include "velox/expression/Expr.h"
+#include "velox/expression/tests/FuzzerToolkit.h"
 #include "velox/vector/VectorSaver.h"
 #include "velox/vector/fuzzer/VectorFuzzer.h"
 
@@ -48,81 +49,6 @@ void logRowVector(const RowVectorPtr& rowVector) {
       LOG(INFO) << "\tAt " << i << ": " << rowVector->toString(i);
     }
   }
-}
-
-bool isInvalidArgumentOrUnsupported(const VeloxException& ex) {
-  return ex.errorCode() == "INVALID_ARGUMENT" ||
-      ex.errorCode() == "UNSUPPORTED";
-}
-
-void compareExceptions(
-    const VeloxException& left,
-    const VeloxException& right) {
-  // Error messages sometimes differ; check at least error codes.
-  // Since the common path may peel the input encoding off, whereas the
-  // simplified path flatten input vectors, the common and the simplified
-  // paths may evaluate the input rows in different orders and hence throw
-  // different exceptions depending on which bad input they come across
-  // first. We have seen this happen for the format_datetime Presto function
-  // that leads to unmatched error codes UNSUPPORTED vs. INVALID_ARGUMENT.
-  // Therefore, we intentionally relax the comparision here.
-  if (left.errorCode() == right.errorCode() ||
-      (isInvalidArgumentOrUnsupported(left) &&
-       isInvalidArgumentOrUnsupported(right))) {
-    VELOX_CHECK_EQ(left.errorSource(), right.errorSource());
-    VELOX_CHECK_EQ(left.exceptionName(), right.exceptionName());
-    if (left.message() != right.message()) {
-      LOG(WARNING) << "Two different VeloxExceptions were thrown:\n\t"
-                   << left.message() << "\nand\n\t" << right.message();
-    }
-  } else {
-    LOG(ERROR) << left.what();
-    LOG(ERROR) << right.what();
-    VELOX_FAIL(
-        "Common path and simplified path threw different exceptions: {} vs. {}",
-        left.message(),
-        right.message());
-  }
-}
-
-void compareExceptions(
-    std::exception_ptr commonPtr,
-    std::exception_ptr simplifiedPtr) {
-  // If we don't have two exceptions, fail.
-  if (!commonPtr || !simplifiedPtr) {
-    if (!commonPtr) {
-      LOG(ERROR) << "Only simplified path threw exception:";
-      std::rethrow_exception(simplifiedPtr);
-    }
-    LOG(ERROR) << "Only common path threw exception:";
-    std::rethrow_exception(commonPtr);
-  }
-
-  // Otherwise, make sure the exceptions are the same.
-  try {
-    std::rethrow_exception(commonPtr);
-  } catch (const VeloxException& ve1) {
-    try {
-      std::rethrow_exception(simplifiedPtr);
-    } catch (const VeloxException& ve2) {
-      compareExceptions(ve1, ve2);
-      return;
-    } catch (const std::exception& e2) {
-      LOG(WARNING) << "Two different exceptions were thrown:\n\t"
-                   << ve1.message() << "\nand\n\t" << e2.what();
-    }
-  } catch (const std::exception& e1) {
-    try {
-      std::rethrow_exception(simplifiedPtr);
-    } catch (const std::exception& e2) {
-      if (e1.what() != e2.what()) {
-        LOG(WARNING) << "Two different std::exceptions were thrown:\n\t"
-                     << e1.what() << "\nand\n\t" << e2.what();
-      }
-      return;
-    }
-  }
-  VELOX_FAIL("Got two incompatible exceptions.");
 }
 
 void compareVectors(const VectorPtr& left, const VectorPtr& right) {
